+++ Makefile
@@ -1,8 +1,11 @@
 build:
 	go build -buildvcs=false -o ./bin/goblockchain 
 
+all:
+	go build ./... 
+
 run: build
 	./bin/goblockchain
 
 test:
-	go test  ./...
+	go test  ./... -race
