+++ network/server_test.go
@@ -50,11 +50,20 @@ func TestNetwork(t *testing.T) {
 	toplgy.connect(nonValidators[1].ID, nonValidators[2].ID)
 
 	n.setTopology(toplgy)
-	n.initializeConnections()
+	//n.connectAll()
 
 	n.startServers()
 
-	n.runFor(5, cancelFunc)
+	time.Sleep(4 * blockTime)
+
+	lateComer := generateNonValidators(t, l, "late")[0]
+	// the transport layer is kinda fucked
+	// shenignans here
+	n.register(lateComer)
+	n.toplgy.connect(lateComer.ID, v.ID)
+	n.initServer(lateComer, v)
+
+	n.runFor(3, cancelFunc)
 
 	servers := make([]*Server, 0)
 	for _, server := range n.Servers {
@@ -70,7 +79,11 @@ func TestNetwork(t *testing.T) {
 		for j := i + 1; j < len(servers); j++ {
 			serverJ := servers[j]
 			hgtJ := serverJ.chain.Height()
-			assert.InDelta(t, hgtI, hgtJ, 1)
+			assert.InDelta(t, hgtI, hgtJ, 1,
+				"heights differ too much between %s (%d) and %s (%d)",
+				serverI.ID, hgtI,
+				serverJ.ID, hgtJ,
+			)
 			common := int(math.Min(float64(hgtI), float64(hgtJ)))
 			hdrI, err := serverI.chain.GetHeader(uint32(common))
 			assert.NoError(t, err)
@@ -86,10 +99,10 @@ func TestNetwork(t *testing.T) {
 
 func (n *network) runFor(nBlocks int, cancel context.CancelFunc) {
 	x := 1.5 * float64(nBlocks)
-	sec := n.blockTime.Seconds()
-	sleepMill := int64(x*sec) * 1000
+	msec := float64(n.blockTime.Milliseconds())
+	sleepMill := x * msec
 	d := time.Duration(sleepMill) * time.Millisecond
-	zap.L().Sugar().Infof("running network for %d seconds", d.Seconds())
+	zap.L().Sugar().Infof("running network for %d milliseconds", d.Milliseconds())
 
 	time.Sleep(d)
 	zap.L().Sugar().Info("TEST CALLING CANCEL")
@@ -104,10 +117,10 @@ func (n *network) addPeer(from, to *Server) {
 		zap.String("from", from.ID),
 		zap.String("to", to.ID),
 	)
-	fromTransport := from.Transports[0]
+	fromTransport := from.PeerTransports[0]
 	require.NotEmpty(n.t, fromTransport)
 
-	toTransport := to.Transports[0]
+	toTransport := to.PeerTransports[0]
 	require.NotEmpty(n.t, toTransport)
 
 	require.NoError(n.t,
@@ -162,34 +175,49 @@ func (n *network) setTopology(t *topology) {
 	n.toplgy = t
 }
 
-func (n *network) initializeConnections() {
+func (n *network) connectAll() {
 	require.NotNil(n.t, n.toplgy, "network has no toplogy")
 	require.NotNil(n.t, n.Servers, "network has no servers")
 
-	for fromId, peerIds := range n.toplgy.connections {
-		fromServer, exists := n.Servers[fromId]
-		require.True(n.t, exists, "server id %s in topology but not in network", fromId)
-		for _, peerId := range peerIds {
-			peerServer, exists := n.Servers[peerId]
-			require.True(n.t, exists, "server id %s in topology but not in network", peerId)
-
-			n.addPeer(fromServer, peerServer)
-		}
+	for _, s := range n.Servers {
+		n.initializeConnections(s)
 	}
+
+}
+
+func (n *network) initializeConnections(s *Server) {
+	fromServer, exists := n.Servers[s.ID]
+	peerIds := n.toplgy.connections[s.ID]
+	require.True(n.t, exists, "server id %s in topology but not in network", fromServer.ID)
+	for _, peerId := range peerIds {
+		peerServer, exists := n.Servers[peerId]
+		require.True(n.t, exists, "server id %s in topology but not in network", peerId)
+		n.addPeer(fromServer, peerServer)
+	}
+}
+
+func (n *network) initServer(s *Server, seeds ...*Server) {
+	n.initializeConnections(s)
+	// seed is the incoming connection to the given server
+	for _, seed := range seeds {
+		n.logger.Sugar().Info("seeding %s from %s", s.ID, seed.ID)
+		n.addPeer(seed, s)
+	}
+
+	n.t.Logf("test network starting %s", s.ID)
+	n.wg.Add(1)
+	go func(s *Server) {
+		defer n.wg.Done()
+		err := s.Start(n.ctx)
+		n.logger.Sugar().Infof("server %s shutdown", s.ID)
+		require.NoError(n.t, err)
+	}(s)
 }
 
 func (n *network) startServers() {
 
-	for id, server := range n.Servers {
-		n.t.Logf("test network starting %s", id)
-		n.wg.Add(1)
-		go func(s *Server) {
-			defer n.wg.Done()
-			err := s.Start(n.ctx)
-			n.logger.Sugar().Infof("server %s shutdown", s.ID)
-			require.NoError(n.t, err)
-
-		}(server)
+	for _, server := range n.Servers {
+		n.initServer(server)
 	}
 }
 
@@ -241,16 +269,16 @@ func mustMakeServer(t *testing.T, opts ServerOpts) *Server {
 func makeValidatorOpts(id string, tr Transport) ServerOpts {
 	privKey := crypto.MustGeneratePrivateKey()
 	return ServerOpts{
-		PrivateKey: privKey,
-		ID:         id,
-		Transports: []Transport{tr},
+		PrivateKey:     privKey,
+		ID:             id,
+		PeerTransports: []Transport{tr},
 	}
 }
 
 func makeNonValidatorOpts(id string, tr Transport) ServerOpts {
 	return ServerOpts{
-		ID:         id,
-		Transports: []Transport{tr},
+		ID:             id,
+		PeerTransports: []Transport{tr},
 	}
 }
