+++ network/transport.go
@@ -10,6 +10,7 @@ type NetAddr string
 // module in Transport
 type Transport interface {
 	Consume() <-chan RPC
+
 	Connect(Transport) error
 	Send(NetAddr, Payload) error
 	Addr() NetAddr
